problem_b/requirements.txt
argon2-cffi==25.1.0
fastapi==0.116.1
httpx==0.28.1
PyJWT==2.10.1
pytest==8.4.2
pytest-cov==7.0.0
uvicorn==0.35.0
